Extract token management to Maps::TokensChannel
