+++ app/javascript/controllers/map_controller.js
@@ -2,19 +2,14 @@ import { Controller } from "stimulus"
 import consumer from "../channels/consumer"
 
 export default class extends Controller {
-  static targets = ["image", "zoomIn", "zoomOut", "token", "tokenContainer", "tokenDrawer", "fogMask", "fogArea"]
+  static targets = ["image", "zoomIn", "zoomOut", "fogMask", "fogArea", "tokenContainer"]
 
   connect() {
     this.mapId = this.element.dataset.mapId
-    this.operatorCode = this.generateOperatorCode()
     this.zoomAmounts = this.imageTarget.dataset.zoomAmounts.split(",").map(parseFloat)
 
     this.setViewportSize()
     this.updateZoomButtons()
-
-    this.tokenTargets.forEach(target => {
-      target.ondragstart = () => null
-    })
 
     this.channel = consumer.subscriptions.create({
       channel: "MapChannel",
@@ -131,67 +126,6 @@ export default class extends Controller {
     )
   }
 
-  stopPropagation(event) {
-    event.stopPropagation()
-  }
-
-  dragOver(event) {
-    this.recordPointerLocation(event.clientX, event.clientY)
-  }
-
-  recordPointerLocation(clientX, clientY) {
-    this.element.dataset.clientX = clientX
-    this.element.dataset.clientY = clientY
-  }
-
-  startMoveToken(event) {
-    event.stopPropagation()
-
-    const { target, clientX, clientY } = event
-    this.recordPointerLocation(event.clientX, event.clientY)
-    const { left, top } = target.getBoundingClientRect()
-
-    const tokenImage = target.getElementsByClassName("token__image")[0]
-    const { width: tokenWidth, height: tokenHeight } = tokenImage.getBoundingClientRect()
-
-    target.dataset.beingDragged = true
-    target.dataset.offsetX = (tokenWidth / 2) - (clientX - left)
-    target.dataset.offsetY = (tokenHeight / 2) - (clientY - top)
-
-    event.dataTransfer.setData("text/plain", target.dataset.tokenId)
-  }
-
-  moveToken(event) {
-    event.stopPropagation()
-
-    const { target } = event
-    const { x: currentX, y: currentY, offsetX, offsetY, tokenId } = target.dataset
-    const { zoomAmount } = this.imageTarget.dataset
-    const { x, y } = this.mapPositionOf(
-      new MouseEvent("drag", {
-        clientX: this.element.dataset.clientX,
-        clientY: this.element.dataset.clientY
-      })
-    )
-
-    const newX = Math.floor(x + (parseFloat(offsetX) / parseFloat(zoomAmount)))
-    const newY = Math.floor(y + (parseFloat(offsetY) / parseFloat(zoomAmount)))
-
-    if (newX != currentX || newY != currentY) {
-      this.setTokenLocation(
-        target,
-        newX,
-        newY
-      )
-    }
-  }
-
-  endMoveToken({ target }) {
-    delete target.dataset.beingDragged
-    delete target.dataset.offsetX
-    delete target.dataset.offsetY
-  }
-
   updateZoomButtons() {
     if (this.hasZoomOutTarget && this.hasZoomInTarget) {
       const zoom = parseInt(this.imageTarget.dataset.zoom)
@@ -202,93 +136,10 @@ export default class extends Controller {
 
   cableReceived(data) {
     switch (data.operation) {
-      case "addToken": {
-        if (this.findToken(data.token_id)) {
-          return
-        }
-        this.tokenContainerTarget.insertAdjacentHTML("beforeend", data.token_html);
-        const token = this.findToken(data.token_id)
-        if (data.stashed) {
-          if (this.hasTokenDrawerTarget) {
-            this.tokenDrawerTarget.appendChild(token)
-          } else {
-            token.remove()
-          }
-        }
-        break;
-      }
       case "addPointer": {
         this.tokenContainerTarget.insertAdjacentHTML("beforeend", data.pointer_html)
         break;
       }
-      case "stashToken": {
-        const token = this.findToken(data.token_id)
-        if (this.hasTokenDrawerTarget) {
-          this.tokenDrawerTarget.appendChild(token)
-        } else {
-          token.remove()
-        }
-        break
-      }
     }
-  }
-
-  findToken(tokenId) {
-    return this.tokenTargets.find(token => token.dataset.tokenId == tokenId)
-  }
-
-  setTokenLocation(token, x, y) {
-    token.dataset.x = x
-    token.dataset.y = y
-  }
-
-  dragOverDrawer(event) {
-    event.preventDefault()
-  }
-
-  dropOnDrawer(event) {
-    event.preventDefault()
-    const tokenId = event.dataTransfer.getData("text/plain")
-
-    const token = this.findToken(tokenId)
-    this.tokenDrawerTarget.appendChild(token)
-
-    this.channel.perform(
-      "stash_token",
-      {
-        map_id: this.mapId,
-        token_id: tokenId
-      }
-    )
-  }
-
-  dragOverMap(event) {
-    event.preventDefault()
-  }
-
-  dropOnMap(event) {
-    event.preventDefault()
-    const tokenId = event.dataTransfer.getData("text/plain")
-
-    const token = this.findToken(tokenId)
-    if (token.parentNode != this.tokenContainerTarget) {
-      if (token.dataset.copyOnPlace != "true") {
-        this.tokenContainerTarget.appendChild(token)
-      }
-
-      this.channel.perform(
-        "place_token",
-        {
-          map_id: this.mapId,
-          token_id: tokenId
-        }
-      )
-    }
-  }
-
-  generateOperatorCode() {
-    return Array
-      .from(window.crypto.getRandomValues(new Uint8Array(32)))
-      .map(c => (c < 16 ? "0" : "") + c.toString(16)).join([]);
   }
 }
