+++ app/jobs/token_broadcast_job.rb
@@ -2,6 +2,6 @@ class TokenBroadcastJob < ApplicationJob
   queue_as :default
 
   def perform(token)
-    MapChannel.add_token(token)
+    Maps::TokensChannel.add_token(token)
   end
 end
