+++ app/views/campaigns/_current_map.html.erb
@@ -2,12 +2,12 @@
   <%= content_tag :div,
     class: "flex-1 flex flex-row h-full",
     data: {
-      controller: "map #{"map--fog" if campaign.current_map.fog_enabled?}",
+      controller: "map map--tokens #{"map--fog" if campaign.current_map.fog_enabled?}",
       "map-id": campaign.current_map.id,
-      action: "dragover->map#dragOver"
+      action: "dragover->map--tokens#dragOver"
     } do %>
     <% if admin %>
-      <div class="current-map__token-drawer token-drawer" data-target="map.tokenDrawer" data-action="dragover->map#dragOverDrawer drop->map#dropOnDrawer">
+      <div class="current-map__token-drawer token-drawer" data-target="map--tokens.drawer" data-action="dragover->map--tokens#dragOverDrawer drop->map--tokens#dropOnDrawer">
         <%= link_to new_campaign_map_token_path(campaign, campaign.current_map), remote: true, title: "New Token", class: "token new-token-link" do %>
           <span class="token__image token__image-add">+</span>
         <% end %>
@@ -20,8 +20,8 @@
       data: {
         controller: "css-var #{"sync-values" if admin}",
         "map-id": campaign.current_map.id,
-        target: "map.image #{"map--fog.image" if campaign.current_map.fog_enabled?}",
-        action: "dragover->map#dragOverMap drop->map#dropOnMap mousedown->map#pointTo #{"mousedown->map--fog#startDrawingFog" if campaign.current_map.fog_enabled? && admin} mousedown->map#moveMap",
+        target: "map.image map--tokens.image #{"map--fog.image" if campaign.current_map.fog_enabled?}",
+        action: "dragover->map--tokens#dragOverMap drop->map--tokens#dropOnMap mousedown->map#pointTo #{"mousedown->map--fog#startDrawingFog" if campaign.current_map.fog_enabled? && admin} mousedown->map#moveMap",
         x: campaign.current_map.x,
         y: campaign.current_map.y,
         zoom: campaign.current_map.zoom,
@@ -39,7 +39,7 @@
         "sync-values-keys": "x y"
       } do %>
       <%= content_tag :div, "", class: "current-map__image", style: background_image(campaign.current_map.image) %>
-      <div class="current-map__token-container" data-target="map.tokenContainer">
+      <div class="current-map__token-container" data-target="map.tokenContainer map--tokens.container">
         <%= render campaign.current_map.tokens.where(stashed: false) %>
       </div>
